Create truncate_mnist (broken).py
The problem is only shared variables can be reset to have new values, but T.dot(W, x) does not return a "shared" tensor variable. On the other hand, if the dot product is cast to shared, the gradient function can no longer recognize "W" as the same "W" when differentiating cross_entropy.